requirements.txt
alembic==1.10.4
cachelib==0.10.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
cryptography==40.0.2
email-validator==2.0.0.post2
Flask==2.2.3
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
mysql-connector==2.2.9
mysql-connector-python==8.0.17
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.3
pypiwin32==223
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.7.1
pywin32==305
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
vboxapi==1.0
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
